added auth test - must fail
